+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renaming modules manually now checks for duplicates and increments suffix if needed
 
 ### Changed
+- `read_module_parameter_enum_str()` doesn't raise TypeError anymore when no enums available. Instead gives Logger Info message and returns values.
 - Renamed `clear_channel()` to `reset_channel()`
 - `print_system_information` now prints enum strings for parameters